update the two factor login flow
