+++ apps/photos/src/pages/two-factor/verify/index.tsx
@@ -2,7 +2,7 @@ import VerifyTwoFactor, {
     VerifyTwoFactorCallback,
 } from 'components/TwoFactor/VerifyForm';
 import router from 'next/router';
-import React, { useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { logoutUser, verifyTwoFactor } from 'services/userService';
 import { PAGES } from 'constants/pages';
 import { User } from 'types/user';
@@ -14,9 +14,6 @@ import FormPaper from 'components/Form/FormPaper';
 import FormTitle from 'components/Form/FormPaper/Title';
 import FormPaperFooter from 'components/Form/FormPaper/Footer';
 import { VerticallyCentered } from 'components/Container';
-import { SESSION_KEYS, getKey } from 'utils/storage/sessionStorage';
-import ComlinkCryptoWorker from 'utils/comlink/ComlinkCryptoWorker';
-import { B64EncryptionResult } from 'types/crypto';
 
 export default function Home() {
     const [sessionID, setSessionID] = useState('');
@@ -50,18 +47,6 @@ export default function Home() {
                 id,
             });
             setData(LS_KEYS.KEY_ATTRIBUTES, keyAttributes);
-            const kekEncryptedAttributes: B64EncryptionResult = getKey(
-                SESSION_KEYS.KEY_ENCRYPTION_KEY
-            );
-            if (kekEncryptedAttributes) {
-                const cryptoWorker = await ComlinkCryptoWorker.getInstance();
-                const kek = await cryptoWorker.decryptB64(
-                    kekEncryptedAttributes.encryptedData,
-                    kekEncryptedAttributes.nonce,
-                    kekEncryptedAttributes.key
-                );
-                console.log('kek', kek);
-            }
 
             router.push(PAGES.CREDENTIALS);
         } catch (e) {
